requirements.txt
django==3.2
djangorestframework
djangorestframework-jwt
channels
channels_redis
pyfcm
